ref: Split the decoder into a separate module
