+++ relay-server/src/extractors/mod.rs
@@ -2,12 +2,14 @@ use actix_web::State;
 
 use crate::service::ServiceState;
 
+mod decoder;
 mod forwarded_for;
 mod request_meta;
 mod shared_payload;
 mod signed_json;
 mod start_time;
 
+pub use self::decoder::*;
 pub use self::forwarded_for::*;
 pub use self::request_meta::*;
 pub use self::shared_payload::*;
